agent/pyproject.toml
[tool.poetry]
name = "agentic-rag-agent"
version = "0.1.0"
description = "12-Step Agentic RAG System with Pydantic AI"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.30.0"
pydantic-ai-slim = {extras = ["ag-ui", "openai"], version = "0.0.14"}
python-dotenv = "^1.0.0"
logfire = "0.49.0"
fastapi = "^0.115.0"
chromadb = "^0.4.0"
tavily-python = "^0.3.0"
httpx = "^0.27.2"
tiktoken = "^0.5.0"
numpy = "^1.24.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
agentic-rag = "src.main:main"
